Move utility function from formats/notebook to lib.
Merge these utility functions with testing utilities
which were previously located in test-utils.
