+++ ipymd/formats/tests/test_notebook.py
@@ -7,7 +7,7 @@
 #------------------------------------------------------------------------------
 
 from ...core.format_manager import format_manager, convert
-from ..notebook import _compare_notebooks
+from ...lib.notebook import _assert_notebooks_equal
 from ...utils.utils import _diff, _show_outputs
 from ._utils import (_test_reader, _test_writer,
                      _exec_test_file, _read_test_file)
@@ -27,7 +27,9 @@ def _test_notebook_writer(basename):
     """Check that (test nb) and (test cells ==> nb) are the same.
     """
     converted, expected = _test_writer(basename, 'notebook')
-    assert _compare_notebooks(converted, expected)
+
+    _assert_notebooks_equal(converted, expected, check_notebook_metadata=False)
+    # assert _compare_notebooks(converted, expected)
 
 
 def _test_notebook_notebook(basename):
@@ -37,7 +39,8 @@ def _test_notebook_notebook(basename):
     cells = convert(contents, from_='notebook')
     converted = convert(cells, to='notebook')
 
-    assert _compare_notebooks(contents, converted)
+    _assert_notebooks_equal(contents, converted, check_notebook_metadata=False)
+    # assert _compare_notebooks(contents, converted)
 
 
 def test_notebook_reader():
