+++ ipymd/lib/notebook.py
@@ -26,9 +26,18 @@ def _stream_output_to_result(output):
         return output
 
 
+def _output_ensure_string(*args):
+    """make sure that strings split up as a list are coerced into a
+    single string. """
+    for output in args:
+        for mime, out in output['data'].items():
+            output['data'][mime] = _ensure_string(out)
+
+
 def _assert_cell_outputs_equal(output_0, output_1, check_metadata=True):
     output_0 = _stream_output_to_result(output_0)
     output_1 = _stream_output_to_result(output_1)
+    _output_ensure_string(output_0, output_1)
     assert output_0['output_type'] == output_1['output_type']
     assert output_0['data'] == output_1['data']
     assert output_0['execution_count'] == output_1['execution_count'] or \
@@ -42,11 +51,36 @@ def _assert_cells_equal(cell_0, cell_1, check_metadata=True):
     assert cell_0['cell_type'] == cell_1['cell_type']
     assert _cell_source(cell_0) == _cell_source(cell_1)
     for output_0, output_1 in zip(_cell_outputs(cell_0), _cell_outputs(cell_1)):
-        _assert_cell_outputs_equal(output_0, output_1, check_metadata=check_metadata)
+        _assert_cell_outputs_equal(output_0, output_1,
+                                   check_metadata=check_metadata)
 
 
-def _assert_notebooks_equal(nb_0, nb_1, check_notebook_metadata=True, check_cell_metadata=True):
+def _assert_notebooks_equal(nb_0, nb_1, check_notebook_metadata=True,
+                            check_cell_metadata=True):
     if check_notebook_metadata:
         assert nb_0['metadata'] == nb_1['metadata']
     for cell_0, cell_1 in zip(nb_0['cells'], nb_1['cells']):
         _assert_cells_equal(cell_0, cell_1, check_metadata=check_cell_metadata)
+
+
+def _cell_input(cell):
+    """Return the input of an ipynb cell."""
+    return _ensure_string(cell.get('source', []))
+
+
+def _cell_output(cell):
+    """Return the output of an ipynb cell."""
+    outputs = cell.get('outputs', [])
+    # Add stdout.
+    stdout = ('\n'.join(_ensure_string(output.get('text', ''))
+                        for output in outputs)).rstrip()
+    # Add text output.
+    text_outputs = []
+    for output in outputs:
+        out = output.get('data', {}).get('text/plain', [])
+        out = _ensure_string(out)
+        # HACK: skip <matplotlib ...> outputs.
+        if out.startswith('<matplotlib'):
+            continue
+        text_outputs.append(out)
+    return stdout + '\n'.join(text_outputs).rstrip()
